pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soren-python-sdk"
version = "1.0.0"
description = "Python SDK for Soren platform integration"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Soren Team"}
]
keywords = ["soren", "sdk", "plugin", "nats"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "nats-py>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/SorenHQ/py-plugin-sdk"
Repository = "https://github.com/SorenHQ/py-plugin-sdk"
Issues = "https://github.com/SorenHQ/py-plugin-sdk/issues"

[tool.setuptools]
# Use find_packages to automatically discover all packages
packages = {find = {}}

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt"]
"pysdk" = ["*.md"]

[tool.setuptools.package-dir]
"" = "."

